Fix the hierarchy of client requests
Deriving from `ClientRequest` over-exposed API to the derived classes.
